Add babel packages and implement ES6 modules
