--- src/modules/toDoTasks.js
@@ -1,36 +1,0 @@
-export default toDoTasks = [
-    {
-      description: 'Sneak peak: To Do list',
-      completed: true,
-      index: 1,
-    },
-    {
-        description: 'Pro-tips about your IDE',
-        completed: false,
-        index: 2,
-    },
-  
-    {
-        description: 'Morning session: Peer-to-peer code review',
-        completed: true,
-        index: 3,
-    },
-  
-    {
-        description: 'Daily standup',
-        completed: true,
-        index: 4,
-    },
-  
-    {
-        description: 'Pro-tips about your IDE',
-        completed: false,
-        index: 5,
-    },
-  
-    {
-        description: 'Get better at seeking help',
-        completed: true,
-        index: 6,
-    },
-  ];
